refactor: extract workspace services from AdminViewFormService
